requirements.txt
Django
ipython
Pillow
psycopg2
tzdata
djangorestframework
pytz
python-dotenv
django-filter
coverage
